refactor: Enums movidos para pasta dedicada.
